refactor: put requests into separate folder
